refactor: Change zoom pane into viewport
* split zoom pane into smaller components as viewport
* rename UseZoomPanHelper Type to Viewport
